bot/requirements.txt
twitchio == 2.4.0
python-dotenv == 0.20.0
firebase-admin == 5.2.0
google-cloud-firestore == 2.6.1
